+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Nym (NYM) Cosmos-based chain with NYX builtin token
 - added: (Ethereum) NYM ERC-20 builtin token
+- added: (BSC) NYM BEP-20 builtin token
 
 ## 4.70.0 (2026-01-16)